+++ src/entry-extractor.ts
@@ -1,13 +1,16 @@
 import {WebhookPayload} from '@actions/github/lib/interfaces'
 
 /** Regex explanation
- *                                 --- Matches Bump or Bumps
- *                                 |     --- Matches any non-whitespace character; matching as a few as possible
- *                                 |     |          --- Matches any non-whitespace character
- *                                 |     |          |          --- Matches any non-whitespace character
- *                                 |     |          |          |
+ *   --- Matches Bump, Bumps, Update or Updates, without capturing it
+ *  |     --- Matches any non-whitespace character; matching as a few as possible
+ *  |     |          --- Matches any non-whitespace character
+ *  |     |          |           --- Matches the text 'requirement ' or nothing, without capturing it
+ *  |     |          |           |                --- Matches any non-whitespace character
+ *  |     |          |           |                |
  */
-const TITLE_REGEX = new RegExp(/Bumps? (\S+?) from (\S*) to (\S*)/)
+const TITLE_REGEX = new RegExp(
+  /(?:Update|Bump)s? (\S+?) (?:requirement )?from (\S*) to (\S*)/
+)
 
 export interface DependabotEntry {
   pullRequestNumber: number